ci: Add ability to wait for specific checks
Sometimes status checks are not reported immediately, for example in the
case where a webhook request is sent and the receiver is responsible for
deciding whether to create it. Just waiting for the overall CI status
does not necessarily guarantee that all eventual checks have been
created and have reported their result.

Add a new commandline flag to wait for a specific check to have a
result. This flag can be given multiple times to wait for multiple
statuses.

There are two types of checks in the world of GitHub: checks and
statuses. We check for both so it doesn't matter when you give the flag
which kind of check you're looking for.