all: Add retry support to github HTTP requests
I've seen one or two cases recently where GH has a bit of a blip, and we
immediately exit. If we have retrying support then we could instead
recover from short outages and carry on waiting.

Poke a `RoundTripper` from `go-retryablehttp` into the stack of
transports. Now we have something which looks like this:

  * Request to GH is made
  * Request goes to go-retryablehttp
  * Start looping here: request goes to httpcache
  * If cached, return the result from the cache
  * If not, make a real request (uses a vanilla standard library
    transport here)
    * If that fails, retry. This will bubble up to go-retryablehttp and
      go around the loop again. After N attempts fail, throw an error.
    * If it succeeds, return the result and populate the cache

Since we now have a kind of complicated stack of `Transport`s involved,
add tests for this, checking that retrying and caching work. These are
the beginnings of a testsuite for this project. 👍