go.mod: Fix `go` format
CodeQL is [still telling us that this file is wrong][warning]. Upon
closer reading of [the docs on the file format][docs], `go 1.22` is not
valid since `1.22` was not a release of Go. We need to specify a release
which existed here, and not a minor version of the language which was
what I thought this was before.

Fix it to `go 1.22.2`.

[docs]: https://tip.golang.org/doc/toolchain#version
[warning]: https://github.com/grafana/wait-for-github/actions/runs/8875865988/job/24366210697