ci: Don't want forever when there are no statuses
When there are no statuses on a reference, [GitHub will report the
combined status as `pending`][docs]. This is because status checks can
be added at any time, and so it's not possible to know if the remote end
just hasn't started its check yet, so they play it safe.

That doesn't work for us, though - we need to exit in a finite amount of
time. Currently we don't. With this change, we'll check if there were
any statuses reported if the combined status is `pending`. If there are,
then we really are pending. If there aren't, we will wait for 5 seconds,
look again and if it's still empty then we will say the commit passed
CI. If something has changed then we return "Unknown" and we'll get
called again in due course to see what happened and report the right
thing next time.

[docs]: https://docs.github.com/en/rest/commits/statuses?apiVersion=2022-11-28#get-the-combined-status-for-a-specific-reference