wait-for-github: Show that pagination isn't used
Update the tests to work with multiple pages. Add a couple of new ones
which show that, if the thing we're looking for is not on the first
page, we don't find it and report the wrong result.

This is a demo before we fix the bug.

Also a couple of incidental tidyups: now we're on 1.22 we don't have the
closure variable capture bug so we can stop working around that. And we
can move `t.Parallel()` into the subtests so they all run in parallel,
rather than the parent.