github: Process all pages of results
We've come across some cases now where there are over 100 statuses. In
that case, as 21247305c86ad1a97fa67b9db4473ef478db4898 shows, we don't
find what we're looking for.

The `go-github` package fully supports pagination, we simply weren't
using it. Here we update to use it and fix the issue.