feat(ci): add release-please and prepare project for versioning
Once we have #164 in place I'd like to release a v1 of this project so that
people can depend on it and use (the SHAs of) our tags to get a stable version,
with Dependabot/Renovate for updates and all that. We can get everything in
place so that once #164 is reviewed and merged we'll be able to merge the PR
from `release-please` to get the first release out.

This is inspired by & mostly taken from the similar work we did in
[`grafana/shared-workflows`][shared-workflows] this week.

[shared-workflows]: https://github.com/grafana/shared-workflows