feat(ci): add `ci list` subcommand to list CI checks
It's currently quite difficult to know what the valid values to pass to
`ci --check=<check>` are. This flag is used if you want to limit
checking to only a subset of your CI. For statuses it's quite easy: it's
the text you see in the web UI. For checks, though, you are shown two
pieces of information: `check suite name / check name`. We expect just
the check name.

This is a new subcommand `ci list` which will look at a ref and print a
table of all the CI checks and their status. This is useful to see the
status at a glance from the terminal, and it also bolds the part which
needs to be given to `--check`. That should make using this easier.

This is reusing the GraphQL work done in
7151b52a53becb98caf329b3c4e9d6af38abb521, and adding a few fields to the
data fetched. Namely, we retrieve the fields which give us the number of
statuses / checks and use those to determine if we have any, instead of
figuring that out while we iterate the list. More importantly we
retrieve the workflow name, if there is one, so we can print out the
check's name in the format described above. And finally, we fetch the
`Status` field too. GitHub uses this alongside the `Conclusion` and we
need to look at both to figure out what state a check run is in, if we
want to know if it's pending or whatnot. (For just knowing if a check
run succeeded or failed, you only need the `Conclusion`.)