feat: add advanced codeql workflow
It appears that the protections for the main branch prevent merging PRs
that come from forks where CodeQL has not run on the commit. However,
setting up CodeQL on the fork would also have that file addition be in
the code PR to the origin here. This change adds the default advanced
config for CodeQL in the hopes that the on:pull_request action enables
these CodeQL checks on the PR itself.