ci: build multiarch images natively, generate SBOMs and provenance
We currently build multiarch images using QEMU. This is slow. These days
GitHub offers native arm64 builders for free for public repositories
like ours, so let's use it.

This is following [the recipe outlined on Docker's website][method],
with a couple of adjustments to generate SBOMs and provenance
attestations so that our images can be verified.

[method]: https://docs.docker.com/build/ci/github-actions/multi-platform/#distribute-build-across-multiple-runners