fix: don't share non-thread-safe ANSITransformer instance
In 7eecb18919b5d1fc98620cbe20d9b2dfc1f3fa8a we introduced support for
title casing CI statuses in our output, using the [`text/transform`]
package.

This was done using a `caser` instanced, which was defined as a shared
(global) `ANSITransformer` instance.

Unfortunately doing this with a global variable opens us up to a race
condition which we are triggering with the testsuite. When tests run
with `t.Parallel()`, as we do, multiple goroutines simultaneously call
`transform.String(caser, ...)` on the single instance. Each modifies the
transformer's internal state concurrently, causing data races which
result in incorrect results and, in the case of the testsuite, failing
tests.

The race detector output showed concurrent read/write access to the same
memory location in `ANSITransformer.Reset()`.

Transformers can be fed small chunks of text at a time, and they should
still be able to transform correctly. Imagine we're title casing a
string "hello world". We could feed "he", "llo w", "orld" in three
calls, and the transformer should be able to output "Hello World"
correctly. This means that it must be _stateful_. The `Reset()` method's
existence implies this too.

Fortunately, fix is straightforward: remove the global variable and
create a new transformer instance for each use. Since `ANSITransformer`
instances are lightweight and creation is fast, this doesn't really
impact performance whilst eliminating the shared mutable state that
caused the race condition.

To better be able to detect such bugs in future, add the `-race` flag to
the Docker build's test step. This ensures the CI will catch similar
concurrency issues in future, rather than them slipping through to be
discovered later. The race detector requires `cgo` to be enabled, so we
set `CGO_ENABLED=1` on just the `go test` command in the Dockerfile.
This also requires a compiler and C libraries, which we install.

[`text/transform`]: https://pkg.go.dev/golang.org/x/text/transform