fix(cli): fix `--log-level` flag being ignored
The `--log-level` flag was being completely ignored - even when explicitly
provided on the command line, the application would always use the default
log level instead of the selected level.

The issue was occurring because `cmd.Generic("log-level")` was returning
`nil` in the `Before` hook, even when the flag was provided. This is due
to differences in parsing of root / command-specific flags in
`urfave/cli` v3.

We've fixed this by implementing a flag action (function that executes
when the flag is set), which configures the logger with the
user-specified level. This required implementing `flag.Value` rather
than `TextUnmarshaler` - again a requirement of `urfave/cli` v3.

The phasing of when things are executed is slightly different now. We
have:

1. `Before` hook: initialises the config
2. `Action` handlers for flags which are set on the commandline
3. `Action` handler for the command itself

Because we were setting up the logger in a `Before` hook, we hadn't
handled the `--log-level` flag yet. So we've updated to store an
`*slog.Logger` instead of using the default logger. This means we can
set the field in the config struct directly at any time.

With this, `--log-level` works properly again.